+++ src/components/Nav/Nav.scss
@@ -142,7 +142,7 @@
               text-decoration: none;
               color: $userMenuColor;
             }
-            //클릭했을 때 title 변화!
+
             .active {
               position: absolute;
               height: 27px;
