+++ src/components/Nav/NavData.js
@@ -1,0 +1,19 @@
+const USER_MENU_LIST = [
+  {
+    id: 1,
+    userMenu: '로그인',
+    movePath: '/',
+  },
+  {
+    id: 2,
+    userMenu: '회원가입',
+    movePath: '/sign',
+  },
+  {
+    id: 3,
+    userMenu: '장바구니',
+    movePath: '/cart',
+  },
+];
+
+export default USER_MENU_LIST;
